Refactor: created ImportExportHelper
