+++ DelvUI/Helpers/ImportExportHelper.cs
@@ -1,0 +1,82 @@
+﻿using Dalamud.Plugin;
+using DelvUI.Config;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.IO.Compression;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DelvUI.Helpers
+{
+    public static class ImportExportHelper
+    {
+
+        public static string Separator = "__IMPORTEXPORTSEPARATOR__";
+
+        public static string CompressAndBase64Encode(string jsonString)
+        {
+            using MemoryStream output = new();
+
+            using (DeflateStream gzip = new(output, CompressionLevel.Fastest))
+            {
+                using StreamWriter writer = new(gzip, Encoding.UTF8);
+                writer.Write(jsonString);
+            }
+
+            return Convert.ToBase64String(output.ToArray());
+        }
+
+        public static string Base64DecodeAndDecompress(string base64String)
+        {
+            var base64EncodedBytes = Convert.FromBase64String(base64String);
+
+            using MemoryStream inputStream = new(base64EncodedBytes);
+            using DeflateStream gzip = new(inputStream, CompressionMode.Decompress);
+            using StreamReader reader = new(gzip, Encoding.UTF8);
+            var decodedString = reader.ReadToEnd();
+
+            return decodedString;
+        }
+
+        public static string GenerateExportString(PluginConfigObject configObject)
+        {
+            var jsonString = JsonConvert.SerializeObject(configObject, Formatting.Indented,
+                new JsonSerializerSettings { TypeNameAssemblyFormatHandling = TypeNameAssemblyFormatHandling.Simple, TypeNameHandling = TypeNameHandling.Objects });
+
+            return CompressAndBase64Encode(jsonString);
+        }
+
+        public static T LoadImportString<T>(string importString) where T : PluginConfigObject
+        {
+            try
+            {
+                var jsonString = Base64DecodeAndDecompress(importString);
+                return JsonConvert.DeserializeObject<T>(jsonString);
+            }
+            catch (Exception ex)
+            {
+                PluginLog.Log(ex.Message + "\n" + ex.StackTrace);
+
+                return default;
+            }
+        }
+
+        public static T LoadImportJson<T>(string jsonString) where T : PluginConfigObject
+        {
+            try
+            {
+                return JsonConvert.DeserializeObject<T>(jsonString);
+            }
+            catch (Exception ex)
+            {
+                PluginLog.Log(ex.Message + "\n" + ex.StackTrace);
+
+                return default;
+            }
+        }
+
+    }
+}
